js/src/xpconnect/tests/mochitest/test_bug502959.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=502959
-->
<head>
  <title>Test for Bug 502959</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=502959">Mozilla Bug 502959</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 502959 **/
var xhr = new XMLHttpRequest();

xhr.onreadystatechange = function() { ok(true, "Able to call the double-wrapped function"); };
var doublewrapped = xhr.onreadystatechange;
ok(doublewrapped.toString().indexOf("wrapped") > 0, "got a double-wrapped object back");

(function () {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    Components.utils.forceGC();
})();

doublewrapped.handleEvent({});

</script>
</pre>
</body>
</html>
